Added quick and dirty find page == home
Added a basic barcode finder. If you enter a partial or full barcode, it
will search the db for wines that match hat barcode.
If one is found, it displays all the details for it.
If several are found, it displays a list of them.

This page is now the homepae for the site.